requirements.txt
Flask==3.0.0
pycryptodomex==3.19.0
pyperclip==1.8.2
